- Generating geometry on checkin and generating it on download both working now - Fixed concurrent use of cache

+++ BimServer/src/org/bimserver/longaction/LongActionManager.java
@@ -119,7 +119,7 @@ public class LongActionManager {
 			}
 		}
 	}
-	
+
 	/*
 	 * Untested method
 	 */
@@ -132,7 +132,7 @@ public class LongActionManager {
 		}
 	}
 
-	public synchronized void remove(Integer actionId) {
+	public synchronized void remove(long actionId) {
 		actions.remove(actionId);
 	}
